Break out functionalty to internal class
